- change action for new post response to reload page and scroll to top
- fixed delete post fetch request, added body and header
- changed input type to passowrd in login handlebars

+++ public/js/edit.js
@@ -1,7 +1,9 @@
 const updateFormHandler = async (event) => {
   event.preventDefault();
   if (event.target.hasAttribute("data-id")) {
+    //grab post id
     const id = event.target.getAttribute("data-id");
+    // Grab title and post content 
     const postContent = document.querySelector(".form-content").value.trim();
     const titletUpdate = document.querySelector(".form-title").value.trim();
     const response = await fetch(`/api/post/${id}`, {
@@ -16,23 +18,36 @@ const updateFormHandler = async (event) => {
     });
 
     if (response.ok) {
-      document.location.replace(`/dashboard`);
+      // replace url 
+      document.location.replace("/dashboard");
     } else {
-      alert("Failed to update project");
+      // console log an error message
+      console.log(response.statusText);
     }
   }
 };
 
 const deletePost = async (event) => {
-  const postId = event.target.getAttribute("data-id");
-  const response = await fetch(`/api/post/${postId}`, {
-    method: "DELETE",
-  });
+  if (event.target.hasAttribute("data-id")) {
+    const postId = event.target.getAttribute("data-id");
+    const response = await fetch(`/api/post/${postId}`, {
+      method: "DELETE",
+      body: JSON.stringify({
+        id: postId,
+      }),
+      headers: {
+        "Content-Type": "application/json",
+      },
+    });
 
-  if (response.ok) {
-    window.location.replace("/dashboard");
-  } else {
-    alert("Failed to delete post");
+    if (response.ok) {
+      console.log(response.ok);
+      // redirect to dashboard
+      document.location.href = "/dashboard";
+    } else {
+      // console log an error message
+      console.log(response.statusText);
+    }
   }
 };
